# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusRAG API",
    "version": "v1"
  },
  "paths": {
...
    "/v1/admin/usage/{tenant_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Usage Summary",
        "operationId": "get_usage_summary_v1_admin_usage__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(day|month)$",
              "default": "day",
              "title": "Period"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/nexusrag__apps__api__response__SuccessEnvelope_UsageSummaryResponse___1"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "BAD_REQUEST",
                    "message": "Bad request"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "Missing or invalid bearer token"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "402": {
            "description": "Quota exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUOTA_EXCEEDED",
                    "message": "Monthly request quota exceeded",
                    "details": {
                      "period": "month",
                      "limit": 10000,
                      "used": 10000,
                      "remaining": 0
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_FORBIDDEN",
                    "message": "Insufficient role for this operation"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "Resource not found"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "IDEMPOTENCY_KEY_CONFLICT",
                    "message": "Idempotency-Key already used with different payload"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "REQUEST_VALIDATION_ERROR",
                    "message": "Validation error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMITED",
                    "message": "Rate limit exceeded",
                    "details": {
                      "scope": "api_key",
                      "route_class": "read",
                      "retry_after_ms": 1200
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "Internal server error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "message": "Service unavailable"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorEnvelope"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "nexusrag__apps__api__response__SuccessEnvelope_UsageSummaryResponse___1": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/nexusrag__apps__api__routes__admin__UsageSummaryResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[UsageSummaryResponse]"
      },
...
      "nexusrag__apps__api__routes__admin__UsageSummaryResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "period_type": {
            "type": "string",
            "title": "Period Type"
          },
          "period_start": {
            "type": "string",
            "title": "Period Start"
          },
          "requests_count": {
            "type": "integer",
            "title": "Requests Count"
          },
          "estimated_tokens_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Tokens Count"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "period_type",
          "period_start",
          "requests_count",
          "estimated_tokens_count"
        ],
        "title": "UsageSummaryResponse"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}